prov/efa: update send paths to use app descriptor if available
Update the send paths so that we use the app descriptor regardless of
whether the cache is enabled or not. This bug was causing the EFA
provider to fall back on bounce buffers adding an unneeded memory copy
to the send paths.

More work is needed here to cleanup the registration paths so that we
reuse code to setup iovs that is duplicated, use the app descriptor in a
few other places, however this patch fixes using the app descriptor in
the long message path where it has the most impact.

Also remove stale comment now that we detect GPUDirect support in the
EFA provider.

Signed-off-by: Robert Wespetal <wesper@amazon.com>