prov/rxm: Fix possible invalid pointer access handling rx errors
In rxm_handle_comp_error, if a posted receive has been canceled,
the buffer is freed.  However, if another error occurs (unlikely),
the code assumes that a recv_entry has been associated with the
completion.  This will not be the case for a receive buffer that
has just been returned by the msg endpoint provider.

For the RXM_RX case, if any error occurs with a receive, we
do not have sufficient data to proprogate the error up further.
The failure at the message endpoint level does not correlate
with a failure of an application posted receive.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>