prov/rxm: Use optimized inject limit
Remove increasing the inject size to the size of the eager
message size.  This results in a non-optimal inject size in
some cases, as it requires performing data copies through
bounce buffers.  Bounce buffers can be useful if memory
registration is required, but if the message provider does
not set FI_MR_LOCAL, then the use of bounce buffers may
negatively impact performance.

This is the case when layering rxm over tcp.  As long as the
app doesn't need to reuse the send buffers immediately, we
can avoid both local registration and data copies by only
making use of the core provider's inject size.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>