prov/rxm: Use eager message size instead of inject size
The eager message size may not be the same as the inject
size.  Exchange the eager size used for posted receive
buffers, not the transmit inject size reported to the
application, when exchanging CM data.

Update other places where inject size is used as a stand
in replacement for eager message size.

This likely doesn't cause an issue in practice because the
inject size is increased to match the eager message size.
However, follow on patches will change this.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>