prov/tcp: Use tclass or env var to disable TCP_NODELAY option
By default, the tcp provider configures sockets with TCP_NODELAY.
This works well for latency sensitive applications.  However,
there may be cases where bandwidth can be improved by not enabling
this option.  Provide an environment variable that allows an
admin to disable the feature.  Also adjust the default to
skip setting the option if a tclass of FI_TC_BULK_DATA is specified.
This tclass is the closest that we have to indicating that high
bandwidth is desired over latency.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>