prov/rxm: Rename msg_ep to rx_ep
In rx_buf, the msg_ep field references the endpoint to post
receive buffers to.  This may be a shared receive context or
a normal endpoint from the message provider.  To avoid
confusion, rename msg_ep to rx_ep to clarify that it is
not necessarily the message endpoint.

Rename and update functions called to post buffers to the
rx context.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>