prov/rxm: Enhance check to enable shared rx
In rxm_info_to_core, we check for the tcp provider in
order to enable the use of srx by default.  This flow
works when called through the fi_getinfo() path.  However
it does not work when called through rxm_endpoint().  The
result is that srx ends up not being enabled unless the
use_srx environment variable has been set.

The cause for this is that rxm_endpoint() has the fi_info
passed in by the application, and does not use the base
fi_info structures maintained by rxm, which are used by
fi_getinfo.

To handle this other flow, we use hints passed in by the
app if there's no base_info present.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>