util/cq: Rename oflow_err_list to aux_queue
The overflow/error list is an auxiliary queue referenced
by the main CQ.  It is used to record error completions
as well as successful completions when the main CQ is full.
Both of these are unexpected, but can happen, for example,
when flushing posted receives.  This frequently occurs
with RxM during cleanup.

The rename is part of a larger effort to clarify and fix
how the main CQ and auxiliary queue are used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>