util/cq: Rename and cleanup util_cq_oflow_err_entry
Rename to util_cq_aux_entry, as these entries are now
written to the auxiliary queue.  As part of the change,
rename the parent_comp field to cq_slot.  Entries in
the auxiliary queue 'belong' at the same location in the
CQ as the main_entry that they point to.  We need to
record that entry as part of the aux_entr, but the entry
itself is not a completion or 'parent'.

Update related portions of the code to align with the
new naming scheme.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>