util/cq: Fix inconsistent use of auxiliary queue
When an error is added to the CQ, a 'blank' entry is added to
the main CQ.  That entry points to the auxiliary queue for
the completion data, and the aux entry refers back to the
empty slot so that completion order can be maintained.

If a new entry is added to the CQ, but the CQ is full, the
new entry is added to the auxiliary queue.  However, the
auxiliary entry in this case refers back to a valid entry
in the CQ.  If the new entry is an error, that means that
the entry on the main CQ may or may not be valid.

The problem gets worse if there are successful completions
added to the aux queue because of CQ overrun, and then an
error entry is added to the aux queue.

To fix this, we simplify the behavior to make it consistent.
If an entry is added to the aux queue, then the main CQ
entry will always be blank.  It's merely a placeholder in
order to maintain ordering.  As a result of this change, we
can never 'fill' the CQ's circular queue.  The last entry
must remain empty in order to reference entries added to
the overflow list.

This requires significant reworking of the util CQ code,
and updates to callers.  The shm provider is updated to use
more of the cq abstraction calls, rather than open coding the
implementation as a result.

This fixes an issue reporting incorrect completion data when
overflow occurs, resulting in a segfault.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>