prov/cxi: Adjust to communication profile changes
Communication profile allocation has been updated with a traffic class
type field. Three types are defined: CXI_TC_TYPE_HRP,
CXI_TC_TYPE_RESTRICTED, and CXI_TC_TYPE_DEFAULT. The provider is updated
to take advantage of the new traffic class types.

One issue with the traffic class types is that a specific type may not
be enabled for a given traffic class. When this occurs, fallback to the
CXI_TC_BEST_EFFORT with the same traffic class type is attempted. If
this second allocation attempt is successful, the CXI_TC_BEST_EFFORT
version of the requested type is remapped to the original traffic class
and traffic class type.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>