util/av: Return requested AV size if current size is 0
If we attempt to get the size of the AV prior to the first
address being added, the return value will be 0.  That can
lead to a division by 0 error in rxm_conn_get_rx_size(),
for example.

Add a check to ofi_av_size(), and if the current size is 0,
return the size that was assigned to the AV through the
av attributes when it was created.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>