prov/cxi: Assign one PID/SEP context
Refactor SEP support. Previously, SEP contexts were assigned separate
PtlTEs of a PID. This limited the number of possible contexts and did
not allow for the use of separate source buffer PtlTEs for each context
when using RPut.

The new implementation assigns a separate PID value to each SEP context.
That allows each context to use a separate PtlTE for source buffers used
in RPut commands. It also supports at least 256 contexts on a single
SEP.

Signed-off-by: Zach Tiffany <zachary.tiffany@hpe.com>