fabtests/multi_recv: Initialize mr_desc for send buffers
The multi-recv test handles its own memory registration,
but does not set mr_desc.  This is passed to fi_post_tx
for send operations.  The result is that mr_desc is NULL,
and not valid for providers expecting a usable descriptor.

This problem shows up with the new direct send mode added
to rxm over verbs.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>