prov/cxi: Use libcxi Domain reserve API
Use the libcxi Domain Reserve API Before creating a domain. This
eliminates a race when 2 scalable EPs are enabled at the same time.

Previously, the provider attempted to allocate PIDs one-by-one for each
SEP. If two SEPs were enabled at the same time, they could be
auto-assigned to consecutive PIDs. That would lead to address conflicts
when a single SEP could not acquire a sufficiently large, consecutive
range of PIDs. This new interface allows the provider to atomically
allocate a range of PIDs to a single SEP.

Signed-off-by: Zach Tiffany <zachary.tiffany@hpe.com>