prov/efa: use rxr_ep_post_ctrl() to post REQ packets
Currently rxr_ep_post_ctrl_or_queue() is used to posting REQ
packets, which will push the tx_entry in tx_pending_list if
-FI_EAGAIN was encountered.

This is wrong if the REQ is RTM and shm provider is used as
lower provider, because queuing will mess up message ordering.

It is unecessary for other cases, because application can
handle -FI_EAGAIN.

To address this issue, this patch uses rxr_ep_post_ctrl() to
post REQ packets.

Signed-off-by: Wei Zhang <wzam@amazon.com>