prov/efa: move efa_atfork_callback install to fabric init
The fork handler install was moved in commit ff72bf48a9 when fixing a
bug where the rdma-core hugepage environment variable was set too late.

However, this causes the fork handler to be installed if an EFA device
is not present or if EFA is not used. We only risk corruption with fork
when it is called after the bounce buffers are allocated, registered,
and pinned, so delay installing this handler to validate that fork can
be called safely until then.

Signed-off-by: Robert Wespetal <wesper@amazon.com>