includ/ofi.h : fix a bug in ofi_get_page_bytes()
The function ofi_get_page_bytes() is expected to return a value
which is multiple of page_size, but it is not doing that currently.

This is because a recent change to function ofi_get_page_end(),
which is used by ofi_get_page_bytes().

Ths patch address the issue by not using ofi_get_page_end() to
calcuated the ending address.

Signed-off-by: Wei Zhang <wzam@amazon.com>