prov/cxi: Move PtlTE state into struct cxip_pte
c_ptlte_state is a property of a PtlTE. This state should be stored in
cxip_pte instead of having each structe which uses a cxip_pte also
embedded a c_ptlte_state state.

When a PtlTE undergoes a state change, the c_ptlte_state field in
cxip_pte will automatically be updated.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>