prov/tcp: Flush outstanding operations when disabling the EP
Currently operations are flushed from the EP only after the
application has called close.  This delays error reporting,
where an application may only be checking the CQ for
completed transfers.  Instead fail all outstanding transfers
when disabling the endpoint.

As a result of this change, there is no longer a need to
flush the tx_rsp_pend_queue separately when disabling the ep.
The flush_pending_xfers function is removed as a result.

The state of any partially sent or received transfers is also
cleared, so that future calls to progress the EP do not attempt
to access canceled operations.

Derived from changes by: Sylvain Didelot <sdidelot@ddn.com>

Signed-off-by: Sean Hefty <sean.hefty@intel.com>