prov/tcp: Do not disable ep on send failure
When a send error occurs on socket, it may still have
data available to be read.  Allow reads to continue by
keeping the endpoint in the connected state until
a receive error occurs.

This avoids a potential issue where we could be in the
middle of reading a message, flush the receive message
from the endpoint, and using the remaining data in the
socket stream as being a new (unexpected) message.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>