prov/efa: post queued packets before creating new packets, fix state issue
We should post any packets that are queued due to RNR before trying to
post additional messages. In addition, transition back to the RXR_TX_REQ
state when leaving RXR_TX_QUEUED_CTRL. That way if an RNR error is
written sending the queued control message we will add it back to the
queued list.

Signed-off-by: Robert Wespetal <wesper@amazon.com>