man: Correct definition of Tx/Rx size to match behavior
The man pages indicate that the tx:size and rx:size fields
are the minimal number of operations that may be posted to
an EP before send/recv will return EAGAIN.  This is not
true, as internal messages may consume some entries, and
EAGAIN can be returned on the first operation while
communication is being setup.

Change the wording, so that it captures the actual behavior
of what the providers are actually doing.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>