prov/tcp: Replace cm_event_type with cm_state
The cm event type is being used to track the connection
state.  Rename from cm_event_type to cm_state to align
with its use.  Also rename the events to reflect the
current state of the endpoint, rather than the next
expected event to generate.

Update the cm state upon sending or receiving a cm message.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>