prov/cxi: Do progress when enough is outstanding
The fi_pingpong utility performs a configurable number of TX operations.
For small message sizes fi_inject() is used. When using inject, the app
does not poll CQs. If a large number of iterations are requested
(fi_pingpong -I 10000), the application will perform injects until the
event queue is overrun. This is not incorrect since injects do not
generate events for the user.

To avoid this, progress EQs during a TX operation when there is a large
amount of TX work outstanding. The threshold for polling is high enough
so that most apps will not trigger it but low enough to prevent
overrunning the EQs.

Signed-off-by: Zach Tiffany <zachary.tiffany@hpe.com>