prov/psm2: Move default uuid from fabric to domain
The calls for creating shared contexts accept fi_tx_attr, which unlike
fi_ep_attr, doesn't have the auth_key. As the result, the default uuid
is used as the job key. Previously the default uuid was stored in the
fabric object. The problem was that the auth_key was also unavailable
when the fabric object is created. Therefore, the default uuid can only
be set via environment variables. This creates discrepancies between
shared contexts and endpoints w.r.t. job key setting. Moving the default
uuid to the domain object solves the issue.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>