prov/shm: retry shm_open without O_EXCL if process is dead
If shm_open returns an error because the file already exists, check
to see if the process is not running anymore and proceed with rewriting
if the process is dead.

This still prevents two processes from using the same shm name while
also allowing for using the same name if the last process was killed with
SIGKILL and did not get cleaned up by the signal handler.

Signed-off-by: aingerson <alexia.ingerson@intel.com>