prov/psm2: Fix a double free issue related to disconnection
In psmx2_av_map_remove(), disconnection call is made without removing
the epaddr from the peer_list and clearing the epaddr context. A later
call to disconnect_all_peers() may cause double free when it tries to
disconnect every epaddr on the peer_list and some of them may have
already been disconnected. The issue is timing dependent because the
disconnection call is asynchronous.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>