prov/cxi: Support SW EP mode only
Software endpoint mode is a feature which requires software to perform
the message matching that hardware would do. The benefit of this mode is
the number of posted receives and unexpected messages can scale past the
hardware LE limit. But, this comes at a performance cost.

Transition from hardware to software endpoint and back is currently not
supported.

FI_CXI_MSG_OFFLOAD is used to enable this feature. Request buffer count
and sizing is controlled with FI_CXI_REQ_BUF_COUNT and
FI_CXI_REQ_BUF_SIZE.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>