util/util_av : do not use huge pages for av_entry_pool
Huge pages allocation requires user to reserve some
memory, which cannot be used by normal allocation.
Therefore, we want to reserve as little huge pages as
possible, and use it only for performance critical pools.

This patch removed the HUGEPAGES flags for av_entry_pool
because it does not fall into that category.

Signed-off-by: Wei Zhang <wzam@amazon.com>