prov/rxm: Replace cq write errors with asserts
Failures to write completions to the CQ propagate backwards,
which in most cases results in trying to write an error
entry to the CQ again.  As a rule, CQ write failures will
be fatal to the application and should result in all
associated endpoints being disabled.

Create wrapper functions to write to the CQ.  Update
callers to use the wrappers, which will include logging,
and provide a small number of locations for a later updates
to handle CQ write errors in a better fashion.  Convert
other calling locations that write entries to the CQ to
print a warning and assert.

This change results in several functions converting from
returning an int to void, which ripples up the call stack
to several functions.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>