prov/efa: exclude EFA when FI_HMEM and FI_LOCAL_COMM are selected
The provider assumes that the SHM provider supports CUDA transfers which
it currently does not, leading to a crash. We don't have a clean way to
fall back to using the device instead at the moment, so exclude EFA when
both of these caps are selected.

NCCL does not require FI_LOCAL_COMM as it will use CUDA IPC for local
communication already, so NCCL can still utilize EFA after this change.

Signed-off-by: Robert Wespetal <wesper@amazon.com>