util/mr_cache: include deferred_cnt/size in when to flush cache
MR cache should be flushed when memory registration count/size
reached hardware limit.

Currently, only cached_cnt/size was used to compare against
hardware limit. This is an issue because the deferred memory
registrations should be included too.

This patch addresses the issue by including deferred_cnt/size
in the condition of when to flush cache.

Signed-off-by: Wei Zhang <wzam@amazon.com>