prov/rxm: Discard completions with invalid states
Several protocol states are invalid and should never occur
as part of completion handling.  These are marked with
asserts, but also return -EBADOPSTATE back from rxm_handle_comp.
Maintain the assert, but do not return an error from
rxm_handle_comp.  Returning an error results in writing
a completion error in the CQ, with an invalid application
context.  If the application tries to process the completion,
it will likely crash.  Discarding the error avoids propogating
the problem up to the application.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>