prov/rxm: Simplify write rendezvous ack setup
RxM will run over either verbs or tcp.  There are optimizations
for direct access to tcp in order to avoid data copies.  For
verbs, it checks if a transfer can fit within the inline size
or not, and uses inline when possible.

The rxm inject path for rendezvous ack messages is therefore
a minor optimization which isn't needed.  Remove these paths
and use the standard paths.  This simplifies the code path and
is preparation for an optimization to complete receive messages
using rendezvous earlier, without waiting for the ack to
complete.

Cleanup long line lengths and restructure the code to improve
readability.  Remove returning failure codes from
rxm_rndv_send_wr_done, which result in writing unaffiliated
completions to the CQ.  Insert assert as these cases are fatal
to the protocol's operation.

Future patches can write error completions for the message
operation that was aborted.  However there's no mechanism to
signal the peer that the operation was aborted.  That too is
left for future work, and likely requires tearing down the
connection to the peer.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>