include/ofi_rbuf, util/cq: rename ofi_cirque_tail, fix aux cq call
Because the naming is confusing, rename ofi_cirque_tail to
ofi_cirque_next and redefine ofi_cirque_tail to be the last committed
entry.

Rename all calls to ofi_cirque_tail to ofi_cirque_free. The only instance
that is not changed is one in the util cq aux queue code path.

The cq aux queue should be adding all aux entries to the last committed
entry in the cirq. The code was calling ofi_cique_tail when it should be
returning the last committed entry.

Also move the commit and check to the beginning of the insert aux function
to make sure that the entry we are adding the aux entry to is always
the last entry in the cq.

Signed-off-by: aingerson <alexia.ingerson@intel.com>