fabtests/unexpected_msg: Add progress after posting sends
The unexpected message test is intended to verify that the
receiver handles tagged sends that will not match with a
posted receive.  However, the structure of the test is such
that this is not guaranted and does not work when run over
tcp.  The problem is that after posting the send operations,
no calls are made to progress the send queue.  As a result,
the sends can sit queued on the send side until an attempt
is made to read the CQ.  By that point, the receives will
have been posted.

Add calls to read the CQs after posting the sends.  This
will help ensure that the sends are at least handled by the
initiator side prior to posting the receive buffers.

Also, change the tag after posting each send.  Post receives
in the opposite order, and complete each receive after posting.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>