prov/cxi: Use netdev for link settings, check AMA
Look up a CXI netdev when initializing the CXI provider. Use the netdev
to discovery link state and speed.

Check for an algorithmic MAC address during initialization. If AMA is
not installed, print a warning and do not allow the interface to be
used.

This allows link state and speed fields to be removed for cxil_info.

Signed-off-by: Zach Tiffany <zachary.tiffany@hpe.com>