prov/rxm: Use eager, sar, and buffer sizes correctly
Users can specify an eager message size through environment
variables.  However, what's not clear to the user is that rxm
silently reserves a portion of that size of internal headers.
Because the eager size may be used elsewhere by the application,
this can result in non-optimal transfers.

For example, when multi-recv buffers are used, the application
can indicate how much free space remains before the buffer
should be released.  MPI sets this equal to the receive buffer
size (16K).  It then uses the multi-recv buffer to receive
transfers up to 16k in size.  However, the actual eager size
is slightly less than 16k, due to the rxm packet header.  The
result is that 16k transfers from MPI end up being transfered
using either the SAR or rendezvous protocols.  This results
in failures, as multi-recv handling isn't prepared to handle
rendezvous messages.

Rework the handling around eager and sar sizes.  Hide all
header space from the application.  Eager size now refers to the
eager space as viewed by the application.  This aligns with the
'buffer size' environment variable.  Maintain a separate size
for the actual buffers used for send and receives.  Update
SAR tracking to align with eager sizes.  Update checks in the
code to ensure that all values are aligned and reasonable.

This simplifies user's expections, as they no longer need to
be concerned with a header size that they have no visibility into.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>