prov/cxi: Always leave PID in the initiator
Prior to this commit, if logical matching was used, PID was scrubbed
from the initiator value immediately. For operations where an unexpected
header was onloaded and the initiating endpoint was an SEP, this PID
value is needed to issue the RGet.

To resolve this issue, always store PID with the initiator and only
remove it when the target event is to be generated.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>