prov/verbs: Modify XRC to use simplex connection model
Add OFI_OPS_SIMPLEX_CM op support to XRC and remove the
previous logic that created duplex connectivity transparently
to RxM. This simplifies XRC connection code maintenance and
allows RxM to control if and when duplex connectivity is made.

With OFI_OPS_SIMPLEX_CM support, and XRC endpoint may be
used in both a connect and accept call.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>