prov/gni: allow NULL desc in read|write{v,msg}
Libfabric allows relaxed enforcement of local MR registration thus the
user can choose to skip local buffer registration when issuing an RMA.
The current code, however, requires a pointer to a NULL pointer for such
a case, which is inconvenient to user. This patch allows a NULL pointer
to be directly passed.

Signed-off-by: Min Si <msi@anl.gov>